+++ danceschool/core/models.py
@@ -1659,7 +1659,7 @@ class PublicEvent(Event):
 
     def __str__(self):
         try:
-            return '%s: %s' % (self.name, getattr(self.eventoccurrence_set.first(),'startTime').strftime('%a., %B %d, %Y, %I:%M %p'))
+            return '%s: %s' % (self.name, self.firstOccurrenceTime.strftime('%a., %B %d, %Y, %I:%M %p'))
         except AttributeError:
             # Event has no occurrences
             return self.name