+++ danceschool/financial/tests.py
@@ -34,13 +34,14 @@ class RevenueTest(DefaultSchoolTestCase):
 
         # Check that association and payment method choices are populated
         self.assertIn(('Cash','Cash'), response.context_data.get('form').fields['paymentMethod'].choices)
-        self.assertIn(3, [x[0] for x in response.context_data.get('form').fields['associateWith'].choices])
 
         # Create a Revenue item that is not associated with a Series/Event for $10
         response = self.client.post(reverse('submitRevenues'),{
+            'grossTotal': 10,
             'total': 10,
+            'adjustments': 0,
+            'fees': 0,
             'category': default_rev_cat.id,
-            'associateWith': 3,
             'description': 'Test Revenue Item',
             'paymentMethod': 'Cash',
             'currentlyHeldBy': self.superuser.id,
@@ -61,9 +62,11 @@ class RevenueTest(DefaultSchoolTestCase):
 
         # Create a second Revenue item that is associated with Series s for $20
         response = self.client.post(reverse('submitRevenues'),{
+            'grossTotal': 20,
             'total': 20,
+            'adjustments': 0,
+            'fees': 0,
             'category': default_rev_cat.id,
-            'associateWith': 2,
             'event': s.id,
             'description': 'Test Associated Revenue Item',
             'paymentMethod': 'Cash',